fix: move internal packages to pkg
