+++ actions/commitguard/main.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/getoutreach/actions/internal/gh"
+	"github.com/getoutreach/actions/pkg/gh"
 	"github.com/google/go-github/v43/github"
 	"github.com/pkg/errors"
 	actions "github.com/sethvargo/go-githubactions"
