+++ actions/opslevel/main.go
@@ -11,9 +11,9 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/getoutreach/actions/internal/gh"
-	"github.com/getoutreach/actions/internal/slack"
+	"github.com/getoutreach/actions/pkg/gh"
 	"github.com/getoutreach/actions/pkg/opslevel"
+	"github.com/getoutreach/actions/pkg/slack"
 	"github.com/google/go-github/v43/github"
 	opslevelGo "github.com/opslevel/opslevel-go/v2022"
 	"github.com/pkg/errors"
