+++ scripts/new-action.sh
@@ -95,7 +95,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/getoutreach/actions/internal/gh"
+	"github.com/getoutreach/actions/pkg/gh"
 	"github.com/google/go-github/v43/github"
 	"github.com/pkg/errors"
 	actions "github.com/sethvargo/go-githubactions"
